+++ tests/test_remove_old_files.py
@@ -2,6 +2,7 @@
 
 import os
 import shutil
+import subprocess
 import sys
 from tempfile import mkdtemp
 
@@ -55,8 +56,8 @@ def test_remove_old_files():
     create_files(['test1', 'test2'])
     assert_files_exist(['test1', 'test2'])
     os.utime('test2', (0, 0))
-    assert os.system(
-        "%s %s --older 100 ." % (sys.executable, test_prog_path)) == 0
+    assert subprocess.call(
+        [sys.executable, test_prog_path, "--older", "100", "."]) == 0
     assert_files_exist('test1')
     assert_files_not_exist('test2')
 
@@ -67,8 +68,8 @@ def test_recursive():
     test4 = os.path.join('subdir', 'test4')
     assert_files_exist([test3, test4])
     os.utime(test4, (0, 0))
-    assert os.system(
-        "%s %s --older 100 ." % (sys.executable, test_prog_path)) == 0
+    assert subprocess.call(
+        [sys.executable, test_prog_path, "--older", "100", "."]) == 0
     assert_files_exist(test3)
     assert_files_not_exist(test4)
 
@@ -80,10 +81,10 @@ def test_remove_empty_directory():
     assert_files_exist([test3, test4])
     os.utime(test3, (0, 0))
     os.utime(test4, (0, 0))
-    assert os.system(
-        "%s %s --older 100 ." % (sys.executable, test_prog_path)) == 0
+    assert subprocess.call(
+        [sys.executable, test_prog_path, "--older", "100", "."]) == 0
     assert_files_exist('subdir')
     assert_files_not_exist([test3, test4])
-    assert os.system(
-        "%s %s -e --older 100 ." % (sys.executable, test_prog_path)) == 0
+    assert subprocess.call(
+        [sys.executable, test_prog_path, "-e", "--older", "100", "."]) == 0
     assert_files_not_exist('subdir')