+++ docs/news.rst
@@ -6,6 +6,8 @@ Version 0.2.0 (2017-05-??)
 
 * Test at Travis and AppVeyor.
 
+* Use subprocess.call() instead of os.system().
+
 0.1.0 (2017-04-16)
 ------------------